chore: refactoring things a little bi
