+++ workspaces/repo-tools/packages/cli/src/commands/workspace/create.ts
@@ -1,9 +1,8 @@
 import inquirer from 'inquirer';
 import chalk from 'chalk';
-import { execSync } from 'child_process';
-import { readFile, writeFile, cp } from 'fs/promises';
-import { join } from 'path';
+
 import { OptionValues } from 'commander';
+import { createWorkspace } from '../../lib/createWorkspace';
 
 export default async (opts: OptionValues) => {
   const answers = await inquirer.prompt([
@@ -31,46 +30,7 @@ export default async (opts: OptionValues) => {
     },
   ]);
 
-  const workspacePath = join('workspaces', answers.name);
-
-  try {
-    const output = execSync(
-      `npx --yes @backstage/create-app --path ${workspacePath} --skip-install`,
-      { input: answers.name },
-    );
-    console.log(output.toString());
-
-    const file = await readFile(join(workspacePath, 'package.json'), {
-      encoding: 'utf8',
-    });
-    const content = JSON.parse(file);
-
-    const additionalDevDependencies = [
-      '@changesets/cli',
-      '@backstage/repo-tools',
-    ];
-
-    content.devDependencies ??= {};
-
-    for (const additionalDependency of additionalDevDependencies) {
-      const version = execSync(
-        `npm show ${additionalDependency} version`,
-      ).toString();
-      content.devDependencies[additionalDependency] = `^${version.trim()}`;
-    }
-
-    content.name = answers.name;
-
-    await writeFile(
-      join(workspacePath, 'package.json'),
-      JSON.stringify(content, null, 2),
-    );
-
-    // Experimental
-    await cp(join(__dirname, '.changeset'), join(workspacePath, '.changeset'), {
-      recursive: true,
-    });
-  } catch (error) {
-    process.exit(1);
-  }
+  await createWorkspace({
+    name: answers.name,
+  });
 };
